+++ src/main/java/shopping/controller/AuthController.java
@@ -1,27 +1,20 @@
 package shopping.controller;
 
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
 import shopping.application.AuthService;
 import shopping.dto.request.LoginRequest;
 import shopping.dto.response.LoginResponse;
 
-@Controller
+@RestController
 public class AuthController {
 
     private final AuthService authService;
 
     public AuthController(AuthService authService) {
         this.authService = authService;
-    }
-
-    @GetMapping("/login")
-    public String loginPage(Model model) {
-        return "login";
     }
 
     @PostMapping("/login/token")
