+++ src/main/java/shopping/controller/CartController.java
@@ -3,20 +3,22 @@ package shopping.controller;
 import java.util.List;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 import shopping.application.CartService;
 import shopping.auth.RequestToken;
 import shopping.dto.request.CartItemCreateRequest;
 import shopping.dto.request.CartItemUpdateRequest;
 import shopping.dto.response.CartItemResponse;
 
-@Controller
+@RestController
+@RequestMapping("/cart/items")
 public class CartController {
 
     private final CartService cartService;
@@ -25,12 +27,8 @@ public class CartController {
         this.cartService = cartService;
     }
 
-    @GetMapping("/cart")
-    public String getCartPage() {
-        return "cart";
-    }
 
-    @PostMapping("/cart/items")
+    @PostMapping
     public ResponseEntity<Void> createCartItem(
             @RequestToken String email,
             @RequestBody CartItemCreateRequest cartItemCreateRequest) {
@@ -39,14 +37,14 @@ public class CartController {
         return ResponseEntity.status(HttpStatus.CREATED).build();
     }
 
-    @GetMapping("/cart/items")
+    @GetMapping
     public ResponseEntity<List<CartItemResponse>> getCartItems(@RequestToken String email) {
 
         List<CartItemResponse> cartItems = cartService.findAllByEmail(email);
         return ResponseEntity.ok().body(cartItems);
     }
 
-    @PatchMapping("/cart/items/{id}")
+    @PatchMapping("/{id}")
     public ResponseEntity<Void> updateCartItemQuantity(
             @RequestToken String email,
             @PathVariable Long id,
@@ -56,7 +54,7 @@ public class CartController {
         return ResponseEntity.ok().build();
     }
 
-    @DeleteMapping("/cart/items/{id}")
+    @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteCartItem(
             @RequestToken String email,
             @PathVariable Long id) {
