+++ src/main/java/shopping/controller/PageController.java
@@ -1,0 +1,35 @@
+package shopping.controller;
+
+import java.util.List;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
+import shopping.application.ProductService;
+import shopping.dto.response.ProductResponse;
+
+@Controller
+public class PageController {
+
+    private final ProductService productService;
+
+    public PageController(ProductService productService) {
+        this.productService = productService;
+    }
+
+    @GetMapping("/")
+    public String index(Model model) {
+        List<ProductResponse> products = productService.findAll();
+        model.addAttribute("products", products);
+        return "index";
+    }
+
+    @GetMapping("/login")
+    public String loginPage(Model model) {
+        return "login";
+    }
+
+    @GetMapping("/cart")
+    public String getCartPage() {
+        return "cart";
+    }
+}
